update models and dtos

+++ Source/GalaFamilyLibrary.FamilyService/DataTransferObjects/FamilyCreationDTO.cs
@@ -4,11 +4,11 @@ public class FamilyCreationDTO
 {
     public string Name { get; set; }
 
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
 
     public ushort Version { get; set; }
 
-    public int UploaderId { get; set; }
+    public long UploaderId { get; set; }
 }
 public class FamilyCallbackCreationDTO : FamilyCreationDTO
 {
